Big changes to memory management
This is the first step towards a safer memory management: stuff allocated by Rust are returned wrapped into a particular kind of object, so that it can later (with some other updates that will follow) be dropped by Rust.

No more pointer-of-pointer getting dereferenced and assigned.

At the moment everything is returned in a slice because otherwise there would be issues with non-typed structs. Hopefully with some extra mind-twisting macros we can remove the `rgb_as_ptr` C macro and directly return ptrs.