Directly return ptrs when possible
Instead of always wrapping objects in a one-element array, directly return the ptr.

The distinction between Box<[T]> and Box<[T; 1]> is needed because the first example is two usizes long, like so:

```
struct box_of_arr {
    T *ptr;
    size_t len;
};
```

while the other is just one usize long, directly with the pointer:

```
struct box_of_one_elem_arr {
    T *ptr
};
```

so we can simplify that and directly read that as a pointer.